Added ability to detect duplicates
Currently storing all the results in a dictionary where the key is the company/chemical name and the value is the link to the nsf page for that company/chemical, I needed some logic to ensure duplicate keys would be included.
To solve this, I appended "(entry _)" where _ is the number of the duplicate entry.